Changed UnitsService to UnitService, its path from units to unit, getAll to list, getOne to retrieve, getOneExtended to retrieveExtended; Updated imports to new path and class; Added docstrings to Unit related services, model;
