+++ web/src/app/shared/services/activity/activity.model.ts
@@ -1,4 +1,4 @@
-import { UnitNo } from '../units/units.model';
+import { UnitNo } from '../unit/unit.model';
 import { UserEmail } from '../user/user.model';
 
 /**
